We're separating the game into three basic concepts: a pool of players, the server, and the view.
The pool of players represent a pool of websocket
connections (which can be manipulated through various sets
of devices) and they will connect to the server on a separate
endpoint ("/game").

The view is nothing but a representation of the game, and it will
receive an update from the server every tick (tick-size not yet
determined, but perhaps 500?). Preferably this will be made using
react and will simply pass a map as props to re-render everything.

Finally the server will listen for websocket connections on *two*
endpoints (/game and /view), and it will send the updated game state
to the /view endpoint each tick and listen for commands on each /game-
connection.