Added custom sounds for suicides and removed ALL legacy code. Should probably remove some of the design stuff as well, but that can be handled later.
Flames still aren't animating properly, but that doesn't seem
to be a big issue. We need to create a proper API the kids can
use, and throttle the output so that it doesn't flood the server.

Might also be an idea to reduce the bomb timer and increase the player speed.

Hopefully movement input can also be added to the game loop, and we can use collision
detection for placing bombs as well as movement.