Promises and all that jazz
So, in a vain attempt to control the data flow of Socketman
(and as a look-ahead for client prediction) I've implemented a
naïve event emitter system server-side which is used by the
Singletons to emit events to the Game class, which in turn
emits events to the game server :dancer:.

Also removed the dummy implementation and instead created a
basic promise-based controller client which will function as a
template for further controls.

I'll create a *workshop* branch which will still have the
todo folders etc. :racehorse:

Man, async programming is the tits.